+++ R/util_classify.R
@@ -101,7 +101,7 @@ util_classify.RasterLayer <- function(x,
 
     if (is.null(weighting)){
       breaks <- .getJenksBreaks(raster::getValues(x), n)
-      x <-  raster::cut(x, breaks=breaks, include.lowest=T)
+      x <-  raster::cut(x, breaks=breaks, include.lowest = TRUE)
     } else {
       x <- .classify(x, weighting)
     }